+++ script/match_method.py
@@ -15,13 +15,13 @@ import csv
 import re
 import pandas as pd
 from tqdm import tqdm
-from funcverbnet.utils import load_tmp
+from funcverbnet.utils import load_tmp, CodeUtil
 
 from funcverbnet.classifier.sentence_classifier import FuncSentenceClassifier
 from funcverbnet.nodes.funcverbnet import FuncVerbNet
 
 
-# from funcverbnet.data_handler.pattern_matcher import PatternMatcher
+from funcverbnet.data_handler.pattern_matcher import PatternMatcher
 
 
 def load_jl(jl_file_path):
@@ -35,9 +35,9 @@ def load_jl(jl_file_path):
 
 
 if __name__ == '__main__':
-    classifier = FuncSentenceClassifier()
-    net = FuncVerbNet()
-    # pattern_matcher = PatternMatcher()
+    # classifier = FuncSentenceClassifier()
+    # net = FuncVerbNet()
+    pattern_matcher = PatternMatcher()
     # reader = pd.read_csv(load_tmp('method_qualified_name_data.csv'), iterator=True, chunksize=1000)
     # for chunk in tqdm(reader):
     #     for index, row in chunk.iterrows():
@@ -48,9 +48,23 @@ if __name__ == '__main__':
     #         #     # print(e, traceback.print_exc())
     #         #     print(e.__class__.__name__, row[1])
     #         data = pattern_matcher.mapping_template_from_qualified_name(row[1])
-    #         # print(data)
+    #         print(data)
 
-    data = load_jl(load_tmp('node_info_3.jl'))
+    text = 'setMd5Hash'
+    text = 'getAsInt'
+    text = 'assertNull'
+    text = 'assertNotNull'
+    text = 'assertTrue'
+    text = 'deleteDirectoryContents'
+    text = 'checkNotEmpty'
+    text = 'notNull'
+    text = 'array'
+    # print(CodeUtil.simplify_method_qualified_name(text))
+    parent, unqualified_name = CodeUtil.simplify_method_qualified_name(text)
+    print(CodeUtil.decamelize_by_substitute_verb(parent, unqualified_name))
+    print(pattern_matcher.mapping_template_from_qualified_name(text))
+
+    # data = load_jl(load_tmp('node_info_3.jl'))
 
     # for item in data:
     #     print(json.dumps(item, indent=4))
@@ -63,30 +77,30 @@ if __name__ == '__main__':
     #         cate_id = classifier.predict(sentence)
     #         print(relation['end_node_info']['category_id'], relation['end_node_info']['category'], cate_id)
 
-    for item in data:
-        # print(json.dumps(item, indent=4))
-        if not item:
-            continue
-        for relation in item['relations']:
-            if 'functionality' not in relation['relation_name']:
-                continue
-            # print(json.dumps(relation, indent=4))
-            if not relation['relation_node']:
-                continue
-            if 'sentence' not in relation['relation_node']['node_info']:
-                continue
-            # print(json.dumps(relation['relation_node'], indent=4))
-            node_info = relation['relation_node']['node_info']
-            sentence = node_info['sentence']
-            sentence = re.compile(r'<[^>]*>|\([^\)]*\)|\[[^\]]*\]|\{[^\}]*\}', re.S).sub('', sentence)
-            cate_id = classifier.predict(sentence)
-            cate = net.find_f_category_by_id(cate_id).name
-            print(sentence, node_info['category_id'], node_info['category'], '#', cate_id, cate)
-            # print(pattern_matcher.mapping_template_copy(sentence))
-            # for sub_relation in relation['relation_node']['relations']:
-            #     if 'functionality' not in sub_relation['relation_name']:
-            #         continue
-            #     print(sub_relation['relation_node'])
+    # for item in data:
+    #     # print(json.dumps(item, indent=4))
+    #     if not item:
+    #         continue
+    #     for relation in item['relations']:
+    #         if 'functionality' not in relation['relation_name']:
+    #             continue
+    #         # print(json.dumps(relation, indent=4))
+    #         if not relation['relation_node']:
+    #             continue
+    #         if 'sentence' not in relation['relation_node']['node_info']:
+    #             continue
+    #         # print(json.dumps(relation['relation_node'], indent=4))
+    #         node_info = relation['relation_node']['node_info']
+    #         sentence = node_info['sentence']
+    #         sentence = re.compile(r'<[^>]*>|\([^\)]*\)|\[[^\]]*\]|\{[^\}]*\}', re.S).sub('', sentence)
+    #         cate_id = classifier.predict(sentence)
+    #         cate = net.find_f_category_by_id(cate_id).name
+    #         print(sentence, node_info['category_id'], node_info['category'], '#', cate_id, cate)
+    #         # print(pattern_matcher.mapping_template_copy(sentence))
+    #         # for sub_relation in relation['relation_node']['relations']:
+    #         #     if 'functionality' not in sub_relation['relation_name']:
+    #         #         continue
+    #         #     print(sub_relation)
 
     # sent_set = set()
     # with open(load_tmp('node.csv'), 'w') as wf: